Refactored components to accept dynamic nextPage props and added a temporary image for cards that do not have images
