+++ src/components/pokemon-cards/pokemon-polaroid/PokemonPolaroid.tsx
@@ -20,12 +20,14 @@ interface PokemonPolaroidProps {
   id: number;
   masterSet: number;
   grandmasterSet: number;
+  nextPage: string;
 }
 
 const PokemonPolaroid: React.FC<PokemonPolaroidProps> = ({
   id,
   masterSet,
   grandmasterSet,
+  nextPage
 }: PokemonPolaroidProps) => {
   const imageUrl = `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${id}.png`;
   const { session } = useAuth();
@@ -83,7 +85,7 @@ const PokemonPolaroid: React.FC<PokemonPolaroidProps> = ({
   return (
     <Link
       href={{
-        pathname: '/filter-cards',
+        pathname: nextPage,
         query: { type: 'pokemon', pId: id },
       }}
       style={{ textDecoration: 'none' }}
